ana sayfa düzenlendi

+++ OrderPayment/Models/CartItem.cs
@@ -11,7 +11,7 @@ namespace OrderPayment.Models
         public int ProductId { get; set; } // Foreign key to identify the product
         public Product Product { get; set; }
 
-        public int Quantity { get; set; } // Quantity of the product in the cart
+        public decimal Quantity { get; set; } // Quantity of the product in the cart
         public decimal Price { get; set; } // Price per unit of the product
         public DateTime AddedAt { get; set; } // Date when the product was added to the cart
     }
